[ch-#157734616]write queries for travis to build
